Documents/Programas/controle-cozinha-ibft/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
gspread
oauth2client
python-multipart
